requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.30.1
azure-identity==1.15.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
camel==0.1.2
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.0.7
cryptography==42.0.5
cycler==0.11.0
dataclasses-json==0.6.1
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.6
distro==1.9.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.104.1
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.10.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.23.4
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jenkspy==0.3.3
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
kubernetes==28.1.0
langchain==0.0.331
langsmith==0.0.60
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.16.1
openai==1.16.0
openpyxl==3.1.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.1
portalocker==2.8.2
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
pypdf==3.17.0
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.34.0
regex==2023.10.3
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scikit-posthocs==0.8.0
scipy==1.11.1
seaborn==0.12.2
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.39.3
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
unidecode==1.3.8
urllib3==1.26.18
uvicorn==0.24.0.post1
virtualenv==20.23.0
wandb==0.15.5
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.6
websocket-client==1.6.4
websockets==12.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0